+++ lib/rubocop/cop/style/alias.rb
@@ -3,10 +3,30 @@
 module RuboCop
   module Cop
     module Style
-      # This cop finds uses of `alias` where `alias_method` would be more
-      # appropriate (or is simply preferred due to configuration), and vice
-      # versa.
-      # It also finds uses of `alias :symbol` rather than `alias bareword`.
+      # This cop enforces the use of either `#alias` or `#alias_method`
+      # depending on configuration.
+      # It also flags uses of `alias :symbol` rather than `alias bareword`.
+      #
+      # @example
+      #
+      #   # EnforcedStyle: prefer_alias
+      #
+      #   # good
+      #   alias bar foo
+      #
+      #   # bad
+      #   alias_method :bar, :foo
+      #   alias :bar :foo
+      #
+      # @example
+      #
+      #   # EnforcedStyle: prefer_alias_method
+      #
+      #   # good
+      #   alias_method :bar, :foo
+      #
+      #   # bad
+      #   alias bar foo
       class Alias < Cop
         include ConfigurableEnforcedStyle
 
@@ -16,34 +36,31 @@ module RuboCop
 
         def on_send(node)
           return unless node.command?(:alias_method)
-          return if style == :prefer_alias_method
-          return if scope_type(node) == :dynamic
+          return unless style == :prefer_alias && alias_keyword_possible?(node)
 
           msg = format(MSG_ALIAS_METHOD, lexical_scope_type(node))
           add_offense(node, :selector, msg)
         end
 
         def on_alias(node)
-          # alias_method can't be used with global variables
-          return if node.each_child_node(:gvar).any?
-          # alias_method can't be used in instance_eval blocks
-          scope_type = scope_type(node)
-          return if scope_type == :instance_eval
+          return unless alias_method_possible?(node)
 
-          if scope_type == :dynamic || style == :prefer_alias_method
+          if scope_type(node) == :dynamic || style == :prefer_alias_method
             add_offense(node, :keyword, MSG_ALIAS)
           elsif node.children.none? { |arg| bareword?(arg) }
             add_offense_for_args(node)
           end
         end
 
-        def add_offense_for_args(node)
-          existing_args  = node.children.map(&:source).join(' ')
-          preferred_args = node.children.map { |a| a.source[1..-1] }.join(' ')
-          arg_ranges     = node.children.map(&:source_range)
-          msg            = format(MSG_SYMBOL_ARGS, preferred_args,
-                                  existing_args)
-          add_offense(node, arg_ranges.reduce(&:join), msg)
+        private
+
+        def alias_keyword_possible?(node)
+          scope_type(node) != :dynamic && node.arguments.all?(&:sym_type?)
+        end
+
+        def alias_method_possible?(node)
+          scope_type(node) != :instance_eval &&
+            node.children.none?(&:gvar_type?)
         end
 
         def autocorrect(node)
@@ -56,7 +73,14 @@ module RuboCop
           end
         end
 
-        private
+        def add_offense_for_args(node)
+          existing_args  = node.children.map(&:source).join(' ')
+          preferred_args = node.children.map { |a| a.source[1..-1] }.join(' ')
+          arg_ranges     = node.children.map(&:source_range)
+          msg            = format(MSG_SYMBOL_ARGS, preferred_args,
+                                  existing_args)
+          add_offense(node, arg_ranges.reduce(&:join), msg)
+        end
 
         # In this expression, will `self` be the same as the innermost enclosing
         # class or module block (:lexical)? Or will it be something else